Update IABot to v2.0beta15
*Now uses CheckIfDead v1.8 which introduces TOR support
*Edits are now marked as minor
*Localizes digits in edit summaries
*Added Persian localization
*A whole lot of bug fixes.
